more work

+++ common/src/main/java/io/github/flemmli97/runecraftory/common/attachment/player/DailyPlayerUpdater.java
@@ -16,7 +16,7 @@ public class DailyPlayerUpdater {
     private int bathCounter;
 
     private final Random random = new Random();
-    private int dailyRandomSeed;
+    private long dailyRandomSeed;
 
     private final PlayerData data;
 
@@ -26,7 +26,7 @@ public class DailyPlayerUpdater {
 
     public void tick(ServerPlayer player) {
         if (this.dailyRandomSeed == 0)
-            this.dailyRandomSeed = player.getRandom().nextInt();
+            this.dailyRandomSeed = player.getRandom().nextLong();
         int day = WorldUtils.day(player.level());
         if (this.lastUpdateDay != day) {
             this.lastUpdateDay = day;
@@ -35,7 +35,7 @@ public class DailyPlayerUpdater {
             this.gaveMonsterItem = false;
             this.ateFood = false;
             this.bathCounter = 0;
-            this.dailyRandomSeed = player.getRandom().nextInt();
+            this.dailyRandomSeed = player.getRandom().nextLong();
         }
     }
 
@@ -65,7 +65,7 @@ public class DailyPlayerUpdater {
         this.bathCounter++;
     }
 
-    public int getDailyRandomSeed() {
+    public long getDailyRandomSeed() {
         return this.dailyRandomSeed;
     }
 
@@ -80,7 +80,7 @@ public class DailyPlayerUpdater {
         compound.putBoolean("GaveMonsterItem", this.gaveMonsterItem);
         compound.putBoolean("AteFood", this.ateFood);
         compound.putInt("BathCounter", this.bathCounter);
-        compound.putInt("DailyRandomSeed", this.dailyRandomSeed);
+        compound.putLong("DailyRandomSeed", this.dailyRandomSeed);
         return compound;
     }
 
@@ -89,6 +89,6 @@ public class DailyPlayerUpdater {
         this.gaveMonsterItem = compound.getBoolean("GaveMonsterItem");
         this.ateFood = compound.getBoolean("AteFood");
         this.bathCounter = compound.getInt("BathCounter");
-        this.dailyRandomSeed = compound.getInt("DailyRandomSeed");
+        this.dailyRandomSeed = compound.getLong("DailyRandomSeed");
     }
 }
